clean nodeconversions
