+++ packages/core/src/api/blockManipulation/commands/replaceBlocks/replaceBlocks.ts
@@ -1,16 +1,15 @@
+import { Node } from "prosemirror-model";
+import { Block, PartialBlock } from "../../../../blocks/defaultBlocks.js";
+import type { BlockNoteEditor } from "../../../../editor/BlockNoteEditor";
 import {
   BlockIdentifier,
   BlockSchema,
   InlineContentSchema,
   StyleSchema,
 } from "../../../../schema/index.js";
-import { Block, PartialBlock } from "../../../../blocks/defaultBlocks.js";
-import type { BlockNoteEditor } from "../../../../editor/BlockNoteEditor";
-import { Node } from "prosemirror-model";
-import {
-  blockToNode,
-  nodeToBlock,
-} from "../../../nodeConversions/nodeConversions.js";
+
+import { blockToNode } from "../../../nodeConversions/blockToNode.js";
+import { nodeToBlock } from "../../../nodeConversions/nodeToBlock.js";
 import { removeBlocksWithCallback } from "../removeBlocks/removeBlocks.js";
 
 export function replaceBlocks<
